Add files via upload
The Card class has getters. toString() needs to be done. 

The Player class has getters, a HasZeroCards method, and a PlayTopCard method.   

PlayerTest is an attempt to create new Players. I'm pretty sure I need a getter for the user's input when they select how many players there are in order to get the right array length. Once we have that getter and the method for the AI picking a category, the AI will be good to go.