Add files via upload
This is my current version of the Database Connection class. James has to add the code for the query to return the average number of draws  and Nana has to add the code for the query to return the largest number of rounds played in a single game. I still need to work on properly establishing the connection, and add some more variables such as query and con(connect).